Use DocumentListener to monitor String @Tunables
The current implementation of the GUI handler for String @Tunables uses ActionListener to monitor the JTextField for changes. However, the JTextField interface only fires ActionEvents when the ENTER key is pressed, so changes may not be recorded if the JPanel containing the tunable is not closed before reading. This patch uses a DocumentListener to update the underlying String any time the JTextField is modified.